Move code to NFTPlayer
